Start moving handshake functions into library. Clean up code.
