+++ implementations/c/test/ockam/handshake/include/handshake_test.h
@@ -1,0 +1,18 @@
+#include <stdlib.h>
+#include "ockam/handshake.h"
+
+#define NAME_SIZE 28
+#define NAME "Noise_XX_25519_AESGCM_SHA256"
+#define MAX_TRANSMIT_SIZE 2048
+#define DHLEN 32
+#define TAG_SIZE 16
+#define VECTOR_SIZE 12
+#define EPI_STRING_SIZE 30
+#define EPI_BYTE_SIZE 15
+#define EPI_INITIATOR "7375626d6172696e6579656c6c6f77"
+#define EPI_RESPONDER "79656c6c6f777375626d6172696e65"
+
+#define INITIATOR_STATIC    "000102030405060708090a0b0c0d0e0f101112131415161718191a1b1c1d1e1f"
+#define RESPONDER_STATIC    "0102030405060708090a0b0c0d0e0f101112131415161718191a1b1c1d1e1f20"
+#define INITIATOR_EPH       "202122232425262728292a2b2c2d2e2f303132333435363738393a3b3c3d3e3f"
+#define RESPONDER_EPH       "4142434445464748494a4b4c4d4e4f505152535455565758595a5b5c5d5e5f60"
