+++ implementations/c/test/ockam/handshake/initiator/config/vault_config.h
