+++ implementations/c/test/ockam/handshake/responder/config/ipaddress.txt
@@ -1,0 +1,2 @@
+192.168.0.77
+8000
