+++ implementations/c/test/ockam/handshake/responder/handshake_responder_test.c
@@ -5,7 +5,7 @@
 #include "ockam/error.h"
 #include "ockam/vault.h"
 #include "ockam/transport.h"
-#include "handshake.h"
+#include "ockam/handshake.h"
 
 OCKAM_VAULT_CFG_s vault_cfg =
 {
@@ -13,232 +13,6 @@ OCKAM_VAULT_CFG_s vault_cfg =
 		.p_host                      = 0,
 		OCKAM_VAULT_EC_CURVE25519
 };
-
-OCKAM_ERR responder_m1_process( HANDSHAKE* p_h, uint8_t* p_m1, uint16_t m1_size )
-{
-	OCKAM_ERR       status = OCKAM_ERR_NONE;
-	uint16_t        m1_offset = 0;
-	uint8_t         key[KEY_SIZE];
-	uint32_t        key_bytes;
-
-	// 1. Pick a static 25519 keypair for this handshake and set it to s
-	string_to_hex( RESPONDER_STATIC, key, &key_bytes );
-	if( KEY_SIZE != key_bytes ) printf("********oh no*********");
-	status = ockam_vault_key_write( OCKAM_VAULT_KEY_STATIC, key, KEY_SIZE );
-	if( OCKAM_ERR_NONE != status ) {
-		log_error(status, "failed to generate static keypair in initiator_step_1");
-		goto exit_block;
-	}
-
-	status = ockam_vault_key_get_pub( OCKAM_VAULT_KEY_STATIC, p_h->s, KEY_SIZE );
-	if( OCKAM_ERR_NONE != status ) {
-		log_error(status, "failed to generate get static public key in initiator_step_1");
-		goto exit_block;
-	}
-
-	// 2. Generate an ephemeral 25519 keypair for this handshake and set it to e
-	string_to_hex( RESPONDER_EPH, key, &key_bytes );
-	if( KEY_SIZE != key_bytes ) printf("********oh no*********");
-	status = ockam_vault_key_write( OCKAM_VAULT_KEY_EPHEMERAL, key, KEY_SIZE );
-	if( OCKAM_ERR_NONE != status ) {
-		log_error(status, "failed to generate static keypair in initiator_step_1");
-		goto exit_block;
-	}
-
-	status = ockam_vault_key_get_pub( OCKAM_VAULT_KEY_EPHEMERAL, p_h->e, KEY_SIZE );
-	if( OCKAM_ERR_NONE != status ) {
-		log_error(status, "failed to generate get static public key in initiator_step_1");
-		goto exit_block;
-	}
-
-	// 3. Set k to empty, Set n to 0
-	p_h->nonce = 0;
-	memset( p_h->k, 0, KEY_SIZE );
-
-	// 4. Set h and ck to 'Noise_XX_25519_AESGCM_SHA256'
-	memset( p_h->h, 0, SHA256_SIZE );
-	memcpy( p_h->h, NAME, NAME_SIZE );
-	memset( p_h->ck, 0, SHA256_SIZE );
-	memcpy( p_h->ck, NAME, NAME_SIZE );
-
-	// 5. h = SHA256(h || prologue),
-	// prologue is empty
-	mix_hash( p_h, NULL, 0 );
-
-	// 6. Read 32 bytes from the incoming message buffer
-	// parse it as a public key, set it to re
-	// h = SHA256(h || re)
-	memcpy( p_h->re, p_m1, KEY_SIZE );
-	m1_offset += KEY_SIZE;
-	print_uint8_str( p_h->re, KEY_SIZE, "\nM1 re: ");
-
-	mix_hash( p_h, p_h->re, KEY_SIZE );
-
-	// h = SHA256( h || payload )
-	status = mix_hash( p_h, NULL, 0 );
-
-exit_block:
-	return status;
-}
-
-OCKAM_ERR responder_m2_make( HANDSHAKE* p_h, uint8_t* p_payload, uint32_t payload_size,
-		uint8_t* p_msg, uint16_t msg_size, uint16_t* p_bytes_written )
-{
-
-	OCKAM_ERR       status = OCKAM_ERR_NONE;
-	uint8_t         cipher_text[MAX_TRANSMIT_SIZE];
-	uint16_t        offset = 0;
-	uint8_t         vector[VECTOR_SIZE];
-
-	// Make sure msg is big enough
-	if( msg_size < payload_size+TAG_SIZE ) {
-		status = OCKAM_ERR_TRANSPORT_BUFFER_TOO_SMALL;
-		goto exit_block;
-	}
-
-	// 1. h = SHA256(h || e.PublicKey),
-	// Write e.PublicKey to outgoing message
-	// buffer, BigEndian
-	status = mix_hash( p_h, p_h->e, KEY_SIZE );
-	if( OCKAM_ERR_NONE != status ) {
-		log_error(status, "failed mix_hash of prologue in initiator_step_1");
-		goto exit_block;
-	}
-	memcpy( p_msg, p_h->e, sizeof(p_h->e) );
-	offset += sizeof(p_h->e);
-
-	// 2. ck, k = HKDF(ck, DH(e, re), 2)
-	// n = 0
-	// secret = ECDH( e, re )
-	status = hkdf_dh( p_h->ck, sizeof(p_h->ck), OCKAM_VAULT_KEY_EPHEMERAL, p_h->re, sizeof(p_h->re), KEY_SIZE, p_h->ck, p_h->k );
-	if( OCKAM_ERR_NONE != status ) {
-		log_error(status, "failed hkdf_dh of prologue in responder_m2_make");
-		goto exit_block;
-	}
-	p_h->nonce = 0;
-	print_uint8_str( p_h->k, KEY_SIZE, "M2 k1:");
-
-	// 3. c = ENCRYPT(k, n++, h, s.PublicKey)
-	// h =  SHA256(h || c),
-	// Write c to outgoing message buffer
-	memset( cipher_text, 0, sizeof(cipher_text) );
-	make_vector( p_h->nonce, vector );
-	status = ockam_vault_aes_gcm_encrypt( p_h->k, KEY_SIZE, vector, sizeof(vector),
-			p_h->h, sizeof(p_h->h), &cipher_text[KEY_SIZE], TAG_SIZE,  p_h->s, KEY_SIZE, cipher_text, KEY_SIZE);
-	if( OCKAM_ERR_NONE != status ) {
-		log_error(status, "failed ockam_vault_aes_gcm_encrypt in responder_m2_make");
-		goto exit_block;
-	}
-	p_h->nonce += 1;
-
-	status = mix_hash( p_h, cipher_text, KEY_SIZE+TAG_SIZE );
-
-	// Copy cypher text into send buffer
-	memcpy( p_msg+offset, cipher_text, KEY_SIZE+TAG_SIZE );
-	offset += KEY_SIZE+TAG_SIZE;
-
-	// 4. ck, k = HKDF(ck, DH(s, re), 2)
-	// n = 0
-
-	status = hkdf_dh( p_h->ck, sizeof(p_h->ck), OCKAM_VAULT_KEY_STATIC, p_h->re, sizeof(p_h->re),
-			KEY_SIZE, p_h->ck, p_h->k );
-	if( OCKAM_ERR_NONE != status ) {
-		log_error(status, "failed hkdf_dh in responder_m2_make");
-		goto exit_block;
-	}
-	print_uint8_str(p_h->k, KEY_SIZE, "M2 k2:");
-	print_uint8_str(p_h->h, SHA256_SIZE, "h");
-	p_h->nonce = 0;
-
-	// 5. c = ENCRYPT(k, n++, h, payload)
-	// h = SHA256(h || c),
-	// payload is empty
-	memset( cipher_text, 0, sizeof(cipher_text) );
-	make_vector( p_h->nonce, vector );
-	print_uint8_str( p_h->k, KEY_SIZE, "M2 encrypt params:\nk: " );
-	print_uint8_str( vector, sizeof(vector), "Vector:");
-	print_uint8_str( p_h->h, SHA256_SIZE, "h:");
-	status = ockam_vault_aes_gcm_encrypt( p_h->k, KEY_SIZE, vector, sizeof(vector),
-			p_h->h, sizeof(p_h->h), &cipher_text[payload_size], TAG_SIZE, NULL, 0, NULL, 0);
-	if( OCKAM_ERR_NONE != status ) {
-		log_error(status, "failed ockam_vault_aes_gcm_encrypt in responder_m2_make");
-		goto exit_block;
-	}
-	print_uint8_str(&cipher_text[payload_size], TAG_SIZE, "-----M2 encrypt2 tag:");
-	p_h->nonce += 1;
-	memcpy( p_msg+offset, cipher_text, payload_size+TAG_SIZE );
-	offset += payload_size + TAG_SIZE;
-	print_uint8_str(cipher_text, TAG_SIZE+payload_size, "TAG");
-	status = mix_hash( p_h, cipher_text, payload_size+TAG_SIZE );
-
-	// Done
-	*p_bytes_written = offset;
-
-exit_block:
-	return status;
-}
-
-OCKAM_ERR responder_m3_process( HANDSHAKE* p_h, uint8_t* p_m3, uint16_t m3_size )
-{
-	printf("\n\n************M3*************\n");
-
-	OCKAM_ERR       status = OCKAM_ERR_NONE;
-	uint8_t         uncipher[MAX_TRANSMIT_SIZE];
-	uint8_t         tag[TAG_SIZE];
-	uint8_t         vector[VECTOR_SIZE];
-	uint32_t        offset = 0;
-
-	// 1. Read 48 bytes the incoming message buffer as c
-	// p = DECRYPT(k, n++, h, c)
-	// h = SHA256(h || c),
-	// parse p as a public key,
-	// set it to rs
-	memset( tag, 0, sizeof(tag) );
-	memcpy( tag, p_m3+offset+KEY_SIZE, TAG_SIZE );
-	make_vector( p_h->nonce, vector );
-	status = ockam_vault_aes_gcm_decrypt( p_h->k, KEY_SIZE, vector, sizeof(vector),
-			p_h->h, sizeof(p_h->h), tag, sizeof(tag), p_m3, KEY_SIZE, uncipher, KEY_SIZE );
-	if( OCKAM_ERR_NONE != status ) {
-		log_error(status, "failed ockam_vault_aes_gcm_decrypt in responder_m3_process");
-		goto exit_block;
-	}
-	memcpy(p_h->rs, uncipher, KEY_SIZE);
-	mix_hash( p_h, p_m3+offset, KEY_SIZE+TAG_SIZE );
-	offset += KEY_SIZE+TAG_SIZE;
-
-	// 2. ck, k = HKDF(ck, DH(e, rs), 2)
-	// n = 0
-	status = hkdf_dh( p_h->ck, sizeof(p_h->ck), OCKAM_VAULT_KEY_EPHEMERAL,
-			p_h->rs, sizeof(p_h->rs), KEY_SIZE, p_h->ck, p_h->k );
-	if( OCKAM_ERR_NONE != status ) {
-		log_error(status, "failed hkdf_dh in responder_m3_process");
-		goto exit_block;
-	}
-	p_h->nonce = 0;
-
-	// 3. Read remaining bytes of incoming message buffer as c
-	// p = DECRYPT(k, n++, h, c)
-	// h = SHA256(h || c),
-	// parse p as a payload,
-	// payload should be empty
-	memset( tag, 0, sizeof(tag) );
-	memcpy( tag, p_m3+offset, TAG_SIZE );
-	make_vector( p_h->nonce, vector );
-	memset(uncipher, 0, sizeof(uncipher));
-	status = ockam_vault_aes_gcm_decrypt( p_h->k, KEY_SIZE, vector, sizeof(vector),
-			p_h->h, sizeof(p_h->h), tag, sizeof(tag), NULL, 0, NULL, 0 );
-	if( OCKAM_ERR_NONE != status ) {
-		log_error(status, "failed ockam_vault_aes_gcm_decrypt in initiator_m2_recv");
-		goto exit_block;
-	}
-	p_h->nonce += 1;
-	print_uint8_str( uncipher, 4, "M3 payload");
-	mix_hash( p_h, p_m3+offset, TAG_SIZE );
-	offset += TAG_SIZE;
-
-exit_block:
-	return status;
-}
 
 OCKAM_ERR responder_epilogue( HANDSHAKE* p_h )
 {
@@ -272,7 +46,7 @@ OCKAM_ERR get_ip_info( OCKAM_INTERNET_ADDRESS* p_address )
 	unsigned    port = 0;
 
 	// Read the IP address to bind to
-	address_file = fopen("../ipaddress.txt", "r");
+	address_file = fopen("../config/ipaddress.txt", "r");
 	if(NULL == address_file) {
 		printf("Create a file called \"ipaddress.txt\" with the IP address to listen on," \
 			"in nnn.nnn.nnn.nnn format and port number\n");
