--- implementations/c/test/ockam/handshake/responder/ipaddress.txt
@@ -1,2 +1,0 @@
-127.0.0.1
-8000
