--- implementations/c/test/ockam/handshake/test_keys.txt
@@ -1,4 +1,0 @@
-init_static=000102030405060708090a0b0c0d0e0f101112131415161718191a1b1c1d1e1f
-resp_static=0102030405060708090a0b0c0d0e0f101112131415161718191a1b1c1d1e1f20
-gen_init_ephemeral=202122232425262728292a2b2c2d2e2f303132333435363738393a3b3c3d3e3f
-gen_resp_ephemeral=4142434445464748494a4b4c4d4e4f505152535455565758595a5b5c5d5e5f60
